Unified SuitManager
We don't need the separate SuitManifestManager. I think the goal was to have the code better separated so it'd be easier to keep everything compartmentalised. But, since everything else in this codebase works with all of the related code mostly in the same place, it's distracting to have "the one thing" being different. At least for now, if our intention is not to rewrite everything.
