+++ Source/Managers/SuitManager.swift
@@ -55,12 +55,19 @@ public class SuitManager: McuManager {
         case uploadResource = 4
     }
     
+    // MARK: Properties
+    
     private var offset: UInt64 = 0
     private var uploadData: Data?
     private var pollAttempts: Int = 0
     private var sessionID: UInt64?
     private var state: SuitManagerState = .none
     private weak var uploadDelegate: SuitManagerDelegate?
+    
+    private var callback: ManifestCallback?
+    private var roleIndex: Int?
+    private var roles: [McuMgrManifestListResponse.Manifest.Role] = []
+    private var responses: [McuMgrManifestStateResponse] = []
     
     // MARK: Init
     
@@ -70,11 +77,62 @@ public class SuitManager: McuManager {
     
     // MARK: List
     
+    public typealias ManifestCallback = ([McuMgrManifestStateResponse], Error?) -> Void
     /**
      Command allows to get information about roles of manifests supported by the device.
      */
-    public func listManifests(callback: @escaping McuMgrCallback<McuMgrManifestListResponse>) {
-        send(op: .read, commandId: SuitID.manifestList, payload: nil, callback: callback)
+    public func listManifest(callback: @escaping ManifestCallback) {
+        self.callback = callback
+        roleIndex = 0
+        roles = []
+        responses = []
+        send(op: .read, commandId: SuitID.manifestList, payload: nil, callback: listManifestCallback)
+    }
+    
+    private func validateNext() {
+        guard let i = roleIndex else { return }
+        if i < roles.count {
+            let role = roles[i]
+            logDelegate?.log("Sending Manifest State command for Role \(role.description)", ofCategory: .suit,
+                                  atLevel: .verbose)
+            getManifestState(for: role, callback: roleStateCallback)
+        } else {
+            callback?(responses, nil)
+        }
+    }
+    
+    // MARK: List Manifest Callback
+    
+    private lazy var listManifestCallback: McuMgrCallback<McuMgrManifestListResponse> = { [weak self] response, error in
+        guard let self else { return }
+        
+        guard error == nil, let response, response.rc != 8 else {
+            self.logDelegate?.log("List Manifest Callback not Supported.", ofCategory: .suit, atLevel: .error)
+            self.callback?([], error)
+            return
+        }
+        
+        let roles = response.manifests.compactMap(\.role)
+        self.roleIndex = 0
+        self.roles = roles
+        if #available(iOS 13.0, *) {
+            let rolesList = ListFormatter.localizedString(byJoining: roles.map(\.description))
+            self.logDelegate?.log("Received Response with Roles: \(rolesList)", ofCategory: .suit, atLevel: .debug)
+        }
+        self.validateNext()
+    }
+    
+    // MARK: Role State Callback
+    
+    private lazy var roleStateCallback: McuMgrCallback<McuMgrManifestStateResponse> = { [weak self] response, error in
+        guard let self else { return }
+        guard error == nil, let response, response.rc != 8 else {
+            self.logDelegate?.log("List Manifest Callback not Supported.", ofCategory: .suit, atLevel: .error)
+            return
+        }
+        self.responses.append(response)
+        self.roleIndex? += 1
+        self.validateNext()
     }
     
     /**
@@ -251,7 +309,7 @@ public class SuitManager: McuManager {
             
             self.sessionID = sessionID
             guard self.uploadDelegate != nil else {
-                self.uploadDelegate?.uploadDidFail(with: SuitUpgradeError.suitDelegateRequiredForResource(resource))
+                self.uploadDelegate?.uploadDidFail(with: SuitManagerError.suitDelegateRequiredForResource(resource))
                 return
             }
             // Stop listening for disconnection since firmware has requested a resource.
@@ -330,6 +388,19 @@ enum SuitManagerState {
             return false
         case .uploadingEnvelope, .uploadingResource:
             return true
+        }
+    }
+}
+
+// MARK: - SuitManagerError
+
+public enum SuitManagerError: Error, LocalizedError {
+    case suitDelegateRequiredForResource(_ resource: FirmwareUpgradeManager.Resource)
+    
+    public var errorDescription: String? {
+        switch self {
+        case .suitDelegateRequiredForResource(let resource):
+            return "A \(String(describing: SuitFirmwareUpgradeDelegate.self)) delegate is required since the firmware is requesting resource \(resource.description)."
         }
     }
 }
